refactor: Move binaries into their own crates
This still should take a while to have full non blocking support to
async and cover it with meaninful tests
