+++ Cargo.toml
@@ -1,57 +1,10 @@
-[package]
-name = "jaysons-cli-utils"
-version = "0.2.1"
-authors = ["Jayson Reis <santosdosreis@gmail.com>"]
-edition = "2018"
-
-[dependencies]
-chrono = "0.4.6"
-clap = "2.32.0"
-color-eyre = "0.5.6"
-csv = "1.0.7"
-env_logger = "0.8.1"
-envfile = "0.2.1"
-flate2 = "1.0.7"
-futures = "0.3.0"
-git2 = "0.13.6"
-http = "0.2.1"
-human_format = "1.0.3"
-lazy_static = "1.3.0"
-log = "0.4.6"
-r2d2 = "0.8.4"
-rayon = "1.0.3"
-regex = "1.4.1"
-reqwest = { version = "0.10.8", features = ["json"] }
-rusoto_core = "0.45.0"
-rusoto_s3 = "0.45.0"
-rusoto_ssm = "0.45.0"
-serde = { version ="1.0.116", features = ["derive"] }
-serde_json = "1.0"
-structopt = "0.3.4"
-tokio = { version = "0.3.0", features = ["rt-multi-thread", "macros", "fs"] }
-tokio-compat-02 = "0.1.0"
-wirefilter-engine = "0.6.1"
-
-[[bin]]
-name = "delete-line"
-path = "src/delete_line.rs"
-
-[[bin]]
-name = "delete-local-branches"
-path = "src/delete_local_branches.rs"
-
-[[bin]]
-name = "aws-flow-logs"
-path = "src/aws_flow_logs.rs"
-
-[[bin]]
-name = "aws-ssm-env-importer"
-path = "src/aws_ssm_env_importer.rs"
-
-[[bin]]
-name = "aws-ssm-env-exporter"
-path = "src/aws_ssm_env_exporter.rs"
-
-[[bin]]
-name = "fix-ksql-deleted-topics"
-path = "src/fix_ksql_deleted_topics.rs"
+[workspace]
+members = [
+    "delete-line",
+    "delete-local-branches",
+    "aws-flow-logs",
+    "aws-ssm-env-importer",
+    "aws-ssm-env-exporter",
+    "fix-ksql-deleted-topics",
+    "aws-logs-utils",
+]
