+++ aws-flow-logs/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "aws-flow-logs"
+version = "0.2.1"
+authors = ["Jayson Reis <santosdosreis@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+aws-logs-utils = { path = "../aws-logs-utils" }
+color-eyre = "0.5.6"
+env_logger = "0.8.1"
+lazy_static = "1.3.0"
+rusoto_core = "0.46.0"
+rusoto_s3 = "0.46.0"
+structopt = "0.3.21"
+tokio = { version = "1.2.0", features = ["rt-multi-thread", "macros"] }
