+++ aws-logs-utils/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "aws-logs-utils"
+version = "0.2.1"
+authors = ["Jayson Reis <santosdosreis@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+aws-s3-utils = { path = "../aws-s3-utils" }
+color-eyre = "0.5.6"
+csv = "1.0.7"
+flate2 = "1.0.7"
+futures = "0.3.0"
+human_format = "1.0.3"
+lazy_static = "1.3.0"
+log = "0.4.6"
+rusoto_s3 = "0.46.0"
+wirefilter-engine = "0.6.1"
+serde = "1.0.124"
+structopt = "0.3.21"
