+++ aws-s3-utils/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "aws-s3-utils"
+version = "0.2.1"
+authors = ["Jayson Reis <santosdosreis@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+eyre = "0.6.5"
+futures = "0.3.0"
+rusoto_core = "0.46.0"
+rusoto_s3 = "0.46.0"
+thiserror = "1.0.24"
