+++ aws-ssm-env-exporter/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "aws-ssm-env-exporter"
+version = "0.2.1"
+authors = ["Jayson Reis <santosdosreis@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+color-eyre = "0.5.6"
+env_logger = "0.8.1"
+envfile = "0.2.1"
+lazy_static = "1.3.0"
+log = "0.4.6"
+regex = "1.4.1"
+rusoto_core = "0.46.0"
+rusoto_ssm = "0.46.0"
+structopt = "0.3.4"
+tokio = { version = "1.2.0", features = ["rt-multi-thread", "macros", "fs"] }
