+++ fix-ksql-deleted-topics/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "fix-ksql-deleted-topics"
+version = "0.2.1"
+authors = ["Jayson Reis <santosdosreis@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+color-eyre = "0.5.6"
+env_logger = "0.8.1"
+http = "0.2.1"
+log = "0.4.6"
+regex = "1.4.1"
+reqwest = { version = "0.11.1", features = ["json"] }
+serde = { version = "1.0.116", features = ["derive"] }
+serde_json = "1.0"
+tokio = { version = "1.2.0", features = ["rt-multi-thread", "macros", "fs"] }
